renamed module to `_spin_operators.py`
reducing verbosity
