+++ src/openfermion/hamiltonians/_spin_operators.py
